Bank of Simba HTMLception part done (2 endpoints, 2 html files)

+++ week-07/day-2/src/main/java/com/greenfoxacademy/bankofsimba/model/BankAccount.java
@@ -27,7 +27,7 @@ public class BankAccount {
     this.name = name;
   }
 
-  public void setBalance(long balance) {
+  public void setBalance(double balance) {
     this.balance = balance;
   }
